+++ tests/vipers-tests/src/lib.rs
@@ -14,6 +14,7 @@ use vipers::*;
 #[error_code]
 pub enum ErrorCode {
     MyError,
+    MyOtherError,
 }
 
 #[account]
@@ -225,5 +226,21 @@ fn test_assert_keys_neq_no_match() {
             )
         },
         ErrorCode::MyError
+    );
+}
+
+#[test]
+fn test_anchor_errors_eq_result() {
+    assert_eq!(
+        (err!(MyError) as Result<()>).into_cmp_error(),
+        (err!(MyError) as Result<()>).into_cmp_error(),
+    );
+}
+
+#[test]
+fn test_anchor_errors_ne_result() {
+    assert_ne!(
+        (err!(MyError) as Result<()>).into_cmp_error(),
+        (err!(MyOtherError) as Result<()>).into_cmp_error(),
     );
 }