+++ vipers/src/error.rs
@@ -29,38 +29,44 @@ pub enum VipersError {
     TokenAccountIsNonZero,
 }
 
-/// Conversions into a [VipersWrappedError].
-pub trait IntoVipersError {
-    /// Converts the value into a [VipersWrappedError].
-    fn into_error(self) -> Option<ComparableError>;
+/// Conversions into a [CmpError].
+pub trait IntoCmpError {
+    /// Converts the value into a [CmpError].
+    fn into_cmp_error(self) -> Option<CmpError>;
 }
 
-impl IntoVipersError for anchor_lang::error::Error {
-    fn into_error(self) -> Option<ComparableError> {
-        Some(ComparableError(self))
+impl<T> IntoCmpError for Result<T> {
+    fn into_cmp_error(self) -> Option<CmpError> {
+        self.err()?.into_cmp_error()
     }
 }
 
-impl IntoVipersError for Option<anchor_lang::error::Error> {
-    fn into_error(self) -> Option<ComparableError> {
-        self?.into_error()
+impl IntoCmpError for anchor_lang::error::Error {
+    fn into_cmp_error(self) -> Option<CmpError> {
+        Some(CmpError(self))
     }
 }
 
-impl From<anchor_lang::error::Error> for ComparableError {
+impl IntoCmpError for Option<anchor_lang::error::Error> {
+    fn into_cmp_error(self) -> Option<CmpError> {
+        self?.into_cmp_error()
+    }
+}
+
+impl From<anchor_lang::error::Error> for CmpError {
     fn from(err: anchor_lang::error::Error) -> Self {
-        ComparableError(err)
+        CmpError(err)
     }
 }
 
-/// An error that can be compared (via equality) to other errors.
+/// A comparable error: an error that can be compared (via equality) to other errors.
 #[repr(transparent)]
 #[derive(Debug)]
-pub struct ComparableError(pub anchor_lang::error::Error);
+pub struct CmpError(pub anchor_lang::error::Error);
 
-impl PartialEq for ComparableError {
+impl PartialEq for CmpError {
     fn eq(&self, other: &Self) -> bool {
-        let (ComparableError(a), ComparableError(b)) = (self, other);
+        let (CmpError(a), CmpError(b)) = (self, other);
         match (a, b) {
             (Error::AnchorError(err_a), Error::AnchorError(err_b)) => {
                 err_a.error_code_number == err_b.error_code_number
@@ -73,9 +79,9 @@ impl PartialEq for ComparableError {
     }
 }
 
-impl Eq for ComparableError {}
+impl Eq for CmpError {}
 
-impl Display for ComparableError {
+impl Display for CmpError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         self.0.fmt(f)
     }
@@ -89,17 +95,19 @@ mod tests {
     fn test_program_errors_equal() {
         assert_eq!(
             anchor_lang::error::Error::ProgramError(ProgramError::InvalidArgument.into())
-                .into_error(),
+                .into_cmp_error(),
             anchor_lang::error::Error::ProgramError(ProgramError::InvalidArgument.into())
-                .into_error()
+                .into_cmp_error()
         );
     }
 
     #[test]
     fn test_program_errors_equal_custom() {
         assert_eq!(
-            anchor_lang::error::Error::ProgramError(ProgramError::Custom(10).into()).into_error(),
-            anchor_lang::error::Error::ProgramError(ProgramError::Custom(10).into()).into_error()
+            anchor_lang::error::Error::ProgramError(ProgramError::Custom(10).into())
+                .into_cmp_error(),
+            anchor_lang::error::Error::ProgramError(ProgramError::Custom(10).into())
+                .into_cmp_error()
         );
     }
 
@@ -107,30 +115,63 @@ mod tests {
     fn test_program_errors_mismatch() {
         assert_ne!(
             anchor_lang::error::Error::ProgramError(ProgramError::InvalidArgument.into())
-                .into_error(),
+                .into_cmp_error(),
             anchor_lang::error::Error::ProgramError(ProgramError::AccountAlreadyInitialized.into())
-                .into_error()
+                .into_cmp_error()
         );
     }
 
     #[test]
     fn test_program_errors_mismatch_custom() {
         assert_ne!(
-            anchor_lang::error::Error::ProgramError(ProgramError::Custom(10).into()).into_error(),
-            anchor_lang::error::Error::ProgramError(ProgramError::Custom(11).into()).into_error()
+            anchor_lang::error::Error::ProgramError(ProgramError::Custom(10).into())
+                .into_cmp_error(),
+            anchor_lang::error::Error::ProgramError(ProgramError::Custom(11).into())
+                .into_cmp_error()
         );
     }
 
     #[test]
     fn test_program_errors_equal_none() {
-        assert_eq!(None.into_error(), None.into_error());
+        assert_eq!(None.into_cmp_error(), None.into_cmp_error());
     }
 
     #[test]
     fn test_program_errors_mismatch_random() {
         assert_ne!(
-            None.into_error(),
-            anchor_lang::error::Error::ProgramError(ProgramError::Custom(11).into()).into_error()
+            None.into_cmp_error(),
+            anchor_lang::error::Error::ProgramError(ProgramError::Custom(11).into())
+                .into_cmp_error()
+        );
+    }
+
+    #[error_code]
+    pub enum ErrorCode {
+        MyError,
+        MyOtherError,
+    }
+
+    #[test]
+    fn test_anchor_errors_eq() {
+        assert_eq!(
+            error!(ErrorCode::MyError).into_cmp_error(),
+            error!(ErrorCode::MyError).into_cmp_error(),
+        );
+    }
+
+    #[test]
+    fn test_anchor_errors_eq_result() {
+        assert_eq!(
+            (err!(ErrorCode::MyError) as Result<()>).into_cmp_error(),
+            (err!(ErrorCode::MyError) as Result<()>).into_cmp_error(),
+        );
+    }
+
+    #[test]
+    fn test_anchor_errors_ne_result() {
+        assert_ne!(
+            (err!(ErrorCode::MyError) as Result<()>).into_cmp_error(),
+            (err!(ErrorCode::MyOtherError) as Result<()>).into_cmp_error(),
         );
     }
 }